:recycle: Move getTransactionHash into transactions module
