+++ src/crypto/hash.js
@@ -12,7 +12,6 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-import { getTransactionBytes } from '../transactions/utils';
 
 /**
  * @method getSha256Hash
@@ -21,6 +20,7 @@ import { getTransactionBytes } from '../transactions/utils';
  *
  * @return {string}
  */
+// eslint-disable-next-line import/prefer-default-export
 export function getSha256Hash(data, format) {
 	if (Buffer.isBuffer(data)) {
 		return Buffer.from(naclInstance.crypto_hash_sha256(data));
@@ -42,16 +42,4 @@ export function getSha256Hash(data, format) {
 	throw new Error(
 		'Unsupported data format. Currently only Buffers or `hex` and `utf8` strings are supported.',
 	);
-}
-
-/**
- * @method getTransactionHash
- * @param transaction Object
- *
- * @return {string}
- */
-
-export function getTransactionHash(transaction) {
-	const bytes = getTransactionBytes(transaction);
-	return getSha256Hash(bytes);
 }
